requirements.txt
# Core ML and Data Science
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
xgboost>=1.6.0
imbalanced-learn>=0.9.0
optuna>=3.0.0

# Backend API
fastapi>=0.95.0
uvicorn[standard]>=0.20.0
pydantic>=1.10.0

# Frontend
streamlit>=1.28.0

# Configuration
PyYAML>=6.0

# Utilities
joblib>=1.2.0
pathlib2>=2.3.0

# Optional: ONNX support
# onnx>=1.12.0
# skl2onnx>=1.12.0

# Development and Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.24.0  # For FastAPI testing

# Code Quality
black>=22.0.0
flake8>=5.0.0
isort>=5.10.0

# Documentation
mkdocs>=1.4.0
mkdocs-material>=8.5.0